+++ src/UniAlteri/States/bootstrap.php
@@ -61,13 +61,6 @@ $finderService = function (DI\ContainerInterface $container) {
 //Register finder generator
 $diContainer->registerService(Loader\FinderInterface::DI_FINDER_SERVICE, $finderService);
 
-//Register injection closure generator
-$injectionClosureService = function () {
-    return new DI\InjectionClosure();
-};
-
-$diContainer->registerService(States\StateInterface::INJECTION_CLOSURE_SERVICE_IDENTIFIER, $injectionClosureService);
-
 //Stated class loader, initialize
 $includePathManager = new Loader\IncludePathManager();
 $loader = new Loader\LoaderStandard($includePathManager);