+++ tests/UniAlteri/Tests/Support/Command/Parser/State/Acme/GoodState/States/StateNormal.php
@@ -145,7 +145,7 @@ class StateNormal implements StateInterface
      * @param string               $scope                 self::VISIBILITY_PUBLIC|self::VISIBILITY_PROTECTED|self::VISIBILITY_PRIVATE
      * @param string|null          $statedClassOriginName
      *
-     * @return DI\InjectionClosureInterface
+     * @return \Closure
      *
      * @throws Exception\MethodNotImplemented is the method does not exist or not available in this scope
      * @throws Exception\InvalidArgument      when the method name is not a string